Goal of Project
MechanicsOfCompositesProject_ECraigie
=====================================

This code was developped as part of a Fall 2014 Mechanics of Composites course taken at McGill University. 

The goal was to develop an interface which received the plybook (Excel) for a laminate along with input off-axis axial and moment loads, and output the stresses, strains and stress factors (safety factors) experienced at the top and bottom of each ply in the laminate. 

From these factors, the critical applied loads required to produce the First Ply Failure (FPF) were determined.